+++ src/component/index.js
@@ -5,3 +5,4 @@ export { Player } from "./player/Player";
 export { SuggestionCard } from "./suggestioncard/SuggestionCard";
 export { RemoveCard } from "./remove card/RemoveCard";
 export { Model } from "./model/Model";
+export { Loader } from "./loader/Loader";
