+++ src/component/loader/Loader.jsx
@@ -1,0 +1,10 @@
+import React from 'react'
+import './loader.css'
+
+export const Loader = () => {
+  return (
+    <div className="loader-conatiner">
+    <span className="loader"></span>
+  </div>
+  )
+}
